src/requirements.txt
ticktick-py
python-dotenv
